Filter self-loops and example entities during extraction (#40)
Three-layer defense against extraction noise:
1. LLM prompt instructs extractor to skip illustrative examples
2. Placeholder name filtering (Alice, Bob, foo, bar, etc.) in is_noisy_entity()
3. Chunk-level example marker detection with confidence dampening (0.5x)

Also rejects self-loop edges where source and target resolve to the same node.

+++ engine/crates/covalence-core/src/consolidation/ontology.rs
@@ -203,7 +203,7 @@ pub async fn build_entity_clusters(
     min_cluster_size: usize,
 ) -> Result<ClusterResult> {
     let rows = sqlx::query_as::<_, (String, i64)>(
-        "SELECT canonical_name, mention_count \
+        "SELECT canonical_name, mention_count::int8 \
          FROM nodes \
          WHERE canonical_name IS NOT NULL \
          GROUP BY canonical_name, mention_count \
